requirements.txt
colorlog
paramiko
pytest-sugar
hexdump
sarge
pyasn1
pyasn1-modules
